+++ tools/gamut_3d_diagrams.py
@@ -34,7 +34,7 @@ color_options = {
     'okhsv': {'chroma': 's', 'lightness': 'v', 'force_bottom': True, 'is_hsv': True},
     'hsi': {'chroma': 's', 'lightness': 'i', 'force_bottom': True},
     'hwb': {'chroma': 'w', 'lightness': 'b', 'force_max_radius': True, 'force_top': True, 'force_bottom': True},
-    'cam16-ucs-jmh': {'chroma': 'm'}
+    'cam16-jmh': {'chroma': 'm'}
 }